Remove the flag `BASE` from `string_utilities.h` (#837)
* Remove the BASE flag.

* Remove locale specific tests.

* Code clean up

* Fix DLL import for MinGW.

* Further fixes for MinGW.

* Simplify the build of unit tests.
