+++ src/input_output/string_utilities.h
@@ -39,121 +39,24 @@ INCLUDES
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*/
 
 #include <string>
-#include <sstream>
-#include <iostream>
 #include <vector>
-#include <stdio.h>
-
-#include "JSBSim_API.h"
 
 /*%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 CLASS DECLARATION
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*/
 
-JSBSIM_API bool is_number(const std::string& str);
+namespace JSBSim {
 JSBSIM_API double atof_locale_c(const std::string& input);
-
-#if !defined(BASE)
-  extern std::string& trim_left(std::string& str);
-  extern std::string& trim_right(std::string& str);
-  extern std::string& trim(std::string& str);
-  extern std::string& trim_all_space(std::string& str);
-  extern std::string& to_upper(std::string& str);
-  extern std::string& to_lower(std::string& str);
-  std::vector <std::string> split(std::string str, char d);
-
-  extern std::string replace(std::string str, const std::string& old, const std::string& newstr);
-#else
-  #include <cctype>
-
-  std::string& trim_left(std::string& str)
-  {
-    while (!str.empty() && isspace((unsigned char)str[0])) {
-      str = str.erase(0,1);
-    }
-    return str;
-  }
-
-  std::string& trim_right(std::string& str)
-  {
-    while (!str.empty() && isspace((unsigned char)str[str.size()-1])) {
-      str = str.erase(str.size()-1,1);
-    }
-    return str;
-  }
-
-  std::string& trim(std::string& str)
-  {
-    if (str.empty()) return str;
-    std::string temp_str = trim_right(str);
-    return str = trim_left(temp_str);
-  }
-
-  std::string& trim_all_space(std::string& str)
-  {
-    for (size_t i=0; i<str.size(); i++) {
-      if (isspace((unsigned char)str[i])) {
-        str = str.erase(i,1);
-        --i;
-      }
-    }
-    return str;
-  }
-
-  std::string& to_upper(std::string& str)
-  {
-    for (size_t i=0; i<str.size(); i++) str[i] = toupper(str[i]);
-    return str;
-  }
-
-  std::string& to_lower(std::string& str)
-  {
-    for (size_t i=0; i<str.size(); i++) str[i] = tolower(str[i]);
-    return str;
-  }
-
-  bool is_number(const std::string& str)
-  {
-    if (str.empty())
-      return false;
-    else
-      return (str.find_first_not_of("+-.0123456789Ee") == std::string::npos);
-  }
-
-  std::vector <std::string> split(std::string str, char d)
-  {
-    std::vector <std::string> str_array;
-    size_t index=0;
-    std::string temp = "";
-
-    trim(str);
-    index = str.find(d);
-    while (index != std::string::npos) {
-      temp = str.substr(0,index);
-      trim(temp);
-      if (!temp.empty()) str_array.push_back(temp);
-      str = str.erase(0,index+1);
-      index = str.find(d);
-    }
-    if (!str.empty()) {
-      temp = trim(str);
-      if (!temp.empty()) str_array.push_back(temp);
-    }
-
-    return str_array;
-  }
-
-  std::string replace(std::string str, const std::string& oldstr, const std::string& newstr)
-  {
-    std::string temp = str;
-    size_t old_idx = str.find(oldstr);
-    if (old_idx != std::string::npos) {
-      temp = str.replace(old_idx, 1, newstr);
-    }
-    return temp;
-  }
-
-#endif
+JSBSIM_API std::string& trim_left(std::string& str);
+JSBSIM_API std::string& trim_right(std::string& str);
+JSBSIM_API std::string& trim(std::string& str);
+JSBSIM_API std::string& trim_all_space(std::string& str);
+JSBSIM_API std::string& to_upper(std::string& str);
+JSBSIM_API std::string& to_lower(std::string& str);
+JSBSIM_API bool is_number(const std::string& str);
+JSBSIM_API std::vector <std::string> split(std::string str, char d);
+JSBSIM_API std::string replace(std::string str, const std::string& old, const std::string& newstr);
+};
 
 //%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
